services/bookings-api/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
SQLAlchemy-Utils==0.37.8
python-dateutil==2.8.2